Data.ExtendedReal

Description

Extension of real numbers with positive/negative infinities (±∞). It is useful for describing various limiting behaviors in mathematics.

Remarks:

∞ - ∞ is left undefined as usual, but we define 0 × ∞ = 0 × -∞ = 0 by following the convention of probability or measure theory.

References:

Wikipedia contributors, "Extended real number line," Wikipedia, The Free Encyclopedia, https://en.wikipedia.org/wiki/Extended_real_number_line (accessed September 1, 2014).

Documentation

data Extended r

Extended r is an extension of r with positive/negative infinity (±∞).

Constructors

NegInf
negative infinity (-∞)

Finite !r
finite value

PosInf
positive infinity (+∞)

inf :: Extended r

Infinity (∞)

isFinite :: Extended r -> Bool

isFinite x = not (isInfinite x).

isInfinite :: Extended r -> Bool

isInfinite x returns True iff x is PosInf or NegInf.